+++ src/Tests/JsonRegression.fs
@@ -5,6 +5,7 @@ open Farmer
 open Farmer.Builders
 open Farmer.Arm
 open System.IO
+open Farmer.ServiceBus
 
 let tests =
     testList "ARM Writer Regression Tests" [
@@ -157,5 +158,25 @@ let tests =
             Expect.equal resource.Name "jsontest" "Account name is wrong"
             Expect.equal resource.Sku.Name "Standard_LRS" "SKU is wrong"
             Expect.equal resource.Kind "StorageV2" "Kind"
+        }
+        
+        test "ServiceBus" {
+            let svcBus = serviceBus {
+                name "farmerbus"
+                sku (ServiceBus.Sku.Premium MessagingUnits.OneUnit)
+                add_queues [ queue { name "queue1" } ]
+                add_topics [
+                    topic {
+                        name "topic1"
+                        add_subscriptions [
+                            subscription {
+                                name "sub1"
+                                add_filters [Rule.CreateCorrelationFilter ("filter1", ["header1", "headervalue1"])]
+                            }
+                        ]
+                    }
+                ]
+            }
+            compareResourcesToJson [ svcBus ] "service-bus.json"
         }
     ]